requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cssselect2==0.7.0
Django==4.1.3
django-admin-interface==0.24.2
django-colorfield==0.8.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.8.1
django-parler==2.3
django-rosetta==0.9.8
Flask==2.2.2
fonttools==4.38.0
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
Pillow==9.3.0
polib==1.1.1
PyAutoGUI==0.9.53
pycparser==2.21
pydyf==0.5.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pyphen==0.13.2
PyRect==0.2.0
PyScreeze==0.1.28
python3-xlib==0.15
pytweening==1.0.4
pywhatkit==5.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tinycss2==1.2.1
urllib3==1.26.13
weasyprint==57.1
webencodings==0.5.1
Werkzeug==2.2.2
wikipedia==1.4.0
zipp==3.11.0
zopfli==0.2.2
